New Version with Features
This version of our page has the subpages features and process. As I was working on it, the subpage "features" was not visible and I didn't dig much deep into this problem as I didn't want to cause some trouble. Prof. Kleinen, can you help me add that subpage? Also we have this gif that's 9mb big, I hope it's fine to have it on our page