New Version of B1 with Features
We added a new subpage to for our project: features. But it was not visible on the preview on hugo. I didn't dig much deep into solving that issue because i don't want to end up causing more trouble. Prof. Kleinen, can you help us add that subpage? We also have a small gif with the size of 9mb, i hope that's fine