B1- infinite-zoom: small changes in context (#400)
CHECKS WERE NOT RUN as the ORIGINAL TITLE WAS INCORRECT.

Also, the list of commits is too long - doesn't match the actual pages to main:

* B1 Infinite zoom

* made it smaller

* New Version of B1 with Features

We added a new subpage to for our project: features. But it was not visible on the preview on hugo. I didn't dig much deep into solving that issue because i don't want to end up causing more trouble. Prof. Kleinen, can you help us add that subpage? We also have a small gif with the size of 9mb, i hope that's fine

* B1 13-07

* B1 fix

* B1 put back smaller images

* Reset project-archive and themes/showtime-theme-2021 to match upstream

* Apply b1-infinitezoom changes

* Update theme to match upstream

* hopefully this time

* set submodules to current versions

* re-add cat for b0

* re-add content/ss24/bachelor/b4-sempy/features.md

* typos and file name convention

* B1 infinite zoom- new version

* no komma

* changed text

---------

Co-authored-by: Barne Kleinen <barne.kleinen@htw-berlin.de>